MAGECLOUD-1252: Symlink var/view_preprocessed
